+++ src/main/java/eu/dissco/virtualcollectionservice/configuration/ApplicationConfiguration.java
@@ -11,8 +11,6 @@ import org.springframework.context.annotation.Configuration;
 @Configuration
 public class ApplicationConfiguration {
 
-  public static final String DATE_STRING = "yyyy-MM-dd'T'HH:mm:ss.SSSXXX";
-
   @Bean
   public ObjectMapper objectMapper() {
     var mapper = new ObjectMapper().findAndRegisterModules();
