+++ src/main/java/eu/dissco/virtualcollectionservice/utils/ApplicationUtils.java
@@ -1,0 +1,9 @@
+package eu.dissco.virtualcollectionservice.utils;
+
+public class ApplicationUtils {
+
+  public static final String DATE_STRING = "yyyy-MM-dd'T'HH:mm:ss.SSSXXX";
+
+  private ApplicationUtils() {
+  }
+}
